+++ CHANGES.rst
@@ -10,6 +10,8 @@
 - Fix an AttributeError accessing the ``preferredCharset`` of an
   ``ICodecTerm`` when no ``ICodecPreferredCharset`` was registered.
 
+- Reach and automatically require 100% test coverage.
+
 2.3.1 (2018-01-09)
 ==================